backend/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5
Flask-JWT-Extended>=4.5.3
Flask-Cors>=4.0.0
Flask-Limiter>=4.1.1
psycopg2-binary>=2.9.9
redis>=5.0.1
google-generativeai>=0.7.0
marshmallow>=3.20.2
gunicorn>=21.2.0
python-dotenv>=1.0.0
requests>=2.31.0
flasgger>=0.9.7
sib-api-v3-sdk>=7.6.0
